requirements.txt
streamlit==1.28.1
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3
openai-whisper==20231117
anthropic==0.7.8
python-dotenv==1.0.0
pydub==0.25.1
